Move board mutations into lib/queries/boards.ts
Reads and writes for the same domain entity now live in one module. The
useBoardPatch primitive and its six wrapper hooks (rename, kind, labels,
voice, stepIds, kidReorderable) join useBoards / useBoard. KidSequence
no longer reaches across the feature boundary into board-builder/ to
import a mutation; it imports from lib/queries/boards alongside its query.

Test moved to lib/queries/boards.mutations.test.tsx; uses the local
import path post-move.
